feat: move all block styles to JS (#185)
