+++ includes/class-core.php
@@ -43,7 +43,6 @@ final class Core {
 		\add_action( 'enqueue_block_editor_assets', [ __CLASS__, 'editor_scripts' ] );
 		\add_filter( 'block_type_metadata', [ __CLASS__, 'block_variations' ] );
 		\add_action( 'init', [ __CLASS__, 'block_pattern_categories' ] );
-		\add_action( 'init', [ __CLASS__, 'register_block_styles' ] );
 	}
 
 	/**
@@ -146,29 +145,6 @@ final class Core {
 			array(
 				'label'       => __( 'Newspack Theme - Post Meta', 'newspack-block-theme' ),
 				'description' => __( 'Patterns bundled with the Newspack Block Theme, specifically built for the post meta.', 'newspack-block-theme' ),
-			)
-		);
-	}
-
-	/**
-	 * Register new block styles.
-	 *
-	 * @since Newspack Block Theme 1.0
-	 */
-	public static function register_block_styles() {
-		register_block_style(
-			'core/column',
-			array(
-				'name'  => 'rounded',
-				'label' => __( 'Rounded', 'newspack-block-theme' ),
-			)
-		);
-
-		register_block_style(
-			'core/list',
-			array(
-				'name'  => 'checked',
-				'label' => __( 'Checked', 'newspack-block-theme' ),
 			)
 		);
 	}
