+++ src/js/editor/block-styles/group.js
@@ -1,0 +1,4 @@
+// Remove the Border style that is coming from Newspack Blocks.
+wp.domReady( () => {
+	wp.blocks.unregisterBlockStyle( 'core/group', 'border' );
+} );
