+++ src/js/editor/block-styles/separator.js
@@ -1,4 +1,4 @@
-// Add style to make the Separator block 'tick' (16px).
+// Add style to make the Separator block 'thick' (16px).
 wp.domReady( () => {
 	wp.blocks.registerBlockStyle( 'core/separator', {
 		name: 'thick',
